Restore python 2.7 support
